refactor(inbox): extract sub-widgets to own files
Extract NotificationGroupTile, SubGroupSection, NotificationTile,
InvitationsView, and LoadMoreButton from inbox_screen.dart into
individual files under lib/features/home/widgets/inbox/.

inbox_screen.dart drops from 628 to 179 lines.
