+++ lib/features/home/widgets/inbox/load_more_button.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:kohera/features/notifications/models/notification_constants.dart';
+
+class LoadMoreButton extends StatelessWidget {
+  const LoadMoreButton({
+    required this.isLoading,
+    required this.onPressed,
+    super.key,
+  });
+
+  final bool isLoading;
+  final VoidCallback onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8),
+      child: Center(
+        child: isLoading
+            ? const Padding(
+                padding: EdgeInsets.all(8),
+                child: SizedBox(
+                  width: 20,
+                  height: 20,
+                  child: CircularProgressIndicator(strokeWidth: 2),
+                ),
+              )
+            : TextButton(
+                onPressed: onPressed,
+                child: const Text(InboxText.loadMore),
+              ),
+      ),
+    );
+  }
+}
