+++ lib/features/home/widgets/inbox_screen.dart
@@ -1,22 +1,16 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
-import 'package:kohera/core/routing/route_names.dart';
 import 'package:kohera/core/services/sub_services/selection_service.dart';
-import 'package:kohera/core/utils/reply_fallback.dart';
-import 'package:kohera/core/utils/time_format.dart';
 import 'package:kohera/features/home/screens/home_shell.dart';
+import 'package:kohera/features/home/widgets/inbox/invitations_view.dart';
+import 'package:kohera/features/home/widgets/inbox/load_more_button.dart';
+import 'package:kohera/features/home/widgets/inbox/notification_group_tile.dart';
 import 'package:kohera/features/home/widgets/mobile_space_drawer.dart';
 import 'package:kohera/features/notifications/enum/inbox_filter.dart';
 import 'package:kohera/features/notifications/models/notification_constants.dart';
-import 'package:kohera/features/notifications/models/notification_group.dart';
-import 'package:kohera/features/notifications/models/thread_sub_group.dart';
 import 'package:kohera/features/notifications/services/inbox_controller.dart';
-import 'package:kohera/features/rooms/widgets/invite_tile.dart';
 import 'package:kohera/shared/widgets/kohera_loader.dart';
-import 'package:kohera/shared/widgets/room_avatar.dart';
-import 'package:matrix/matrix.dart' as matrix_sdk;
 import 'package:provider/provider.dart';
 
 class InboxScreen extends StatefulWidget {
@@ -110,7 +104,7 @@ class _InboxScreenState extends State<InboxScreen> {
           // ── Content ──
           Expanded(
             child: controller.filter == InboxFilter.invitations
-                ? _InvitationsView(cs: cs, tt: tt)
+                ? InvitationsView(cs: cs, tt: tt)
                 : controller.isLoading && controller.grouped.isEmpty
                     ? const Center(child: KoheraLoader())
                     : controller.error != null && controller.grouped.isEmpty
@@ -165,12 +159,12 @@ class _InboxScreenState extends State<InboxScreen> {
                                       (controller.hasMore ? 1 : 0),
                                   itemBuilder: (context, i) {
                                     if (i == controller.grouped.length) {
-                                      return _LoadMoreButton(
+                                      return LoadMoreButton(
                                         isLoading: controller.isLoading,
                                         onPressed: controller.loadMore,
                                       );
                                     }
-                                    return _NotificationGroupTile(
+                                    return NotificationGroupTile(
                                       group: controller.grouped[i],
                                       controller: controller,
                                     );
@@ -179,449 +173,6 @@ class _InboxScreenState extends State<InboxScreen> {
                               ),
           ),
         ],
-      ),
-    );
-  }
-}
-
-// ── Notification group tile ──────────────────────────────────
-class _NotificationGroupTile extends StatelessWidget {
-  const _NotificationGroupTile({
-    required this.group,
-    required this.controller,
-  });
-
-  final NotificationGroup group;
-  final InboxController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-    final tt = Theme.of(context).textTheme;
-    final client = controller.client;
-    final room = client.getRoomById(group.roomId);
-
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 8),
-      child: Card(
-        elevation: 0,
-        color: cs.surfaceContainerLow,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(14),
-        ),
-        clipBehavior: Clip.antiAlias,
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // ── Group header ──
-            Padding(
-              padding: const EdgeInsets.fromLTRB(12, 10, 8, 6),
-              child: Row(
-                children: [
-                  if (room != null) ...[
-                    RoomAvatarWidget(room: room, size: 32),
-                    const SizedBox(width: 10),
-                  ],
-                  Expanded(
-                    child: Text(
-                      group.roomName,
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                      style: tt.titleSmall?.copyWith(
-                        fontWeight: FontWeight.w600,
-                      ),
-                    ),
-                  ),
-                  IconButton(
-                    icon: const Icon(Icons.done_all_rounded, size: 20),
-                    tooltip: InboxText.tooltipMarkAsRead,
-                    onPressed: () => controller.markRoomAsRead(group.roomId),
-                    visualDensity: VisualDensity.compact,
-                  ),
-                  IconButton(
-                    icon: const Icon(Icons.open_in_new_rounded, size: 20),
-                    tooltip: InboxText.tooltipOpen,
-                    onPressed: () {
-                      final singleThread = group.subGroups.length == 1
-                          ? group.subGroups.first.threadRootId
-                          : null;
-                      if (singleThread != null) {
-                        context.goNamed(
-                          Routes.roomThread,
-                          pathParameters: {
-                            RouteParams.roomId: group.roomId,
-                            RouteParams.eventId: singleThread,
-                          },
-                        );
-                      } else {
-                        context.goNamed(
-                          Routes.room,
-                          pathParameters: {RouteParams.roomId: group.roomId},
-                        );
-                      }
-                    },
-                    visualDensity: VisualDensity.compact,
-                  ),
-                ],
-              ),
-            ),
-
-            // ── Sub-groups (per-thread + main) ──
-            for (final sub in group.subGroups)
-              _SubGroupSection(
-                roomId: group.roomId,
-                subGroup: sub,
-                client: client,
-                controller: controller,
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-// ── Sub-group section (per-thread or main timeline) ──────────
-class _SubGroupSection extends StatefulWidget {
-  const _SubGroupSection({
-    required this.roomId,
-    required this.subGroup,
-    required this.client,
-    required this.controller,
-  });
-
-  final String roomId;
-  final ThreadSubGroup subGroup;
-  final matrix_sdk.Client client;
-  final InboxController controller;
-
-  @override
-  State<_SubGroupSection> createState() => _SubGroupSectionState();
-}
-
-class _SubGroupSectionState extends State<_SubGroupSection> {
-  String? _rootPreview;
-
-  @override
-  void initState() {
-    super.initState();
-    final id = widget.subGroup.threadRootId;
-    if (id != null) {
-      _rootPreview = widget.controller.rootPreviewFor(id);
-      if (_rootPreview == null) unawaited(_loadRoot(id));
-    }
-  }
-
-  Future<void> _loadRoot(String eventId) async {
-    final room = widget.client.getRoomById(widget.roomId);
-    if (room == null) return;
-    try {
-      final event = await room.getEventById(eventId);
-      if (event == null || !mounted) return;
-      final body = stripReplyFallback(event.body).trim();
-      if (body.isEmpty) return;
-      final truncated = body.length > 80 ? '${body.substring(0, 80)}…' : body;
-      final preview = InboxText.inReplyTo(truncated);
-      widget.controller.setRootPreview(eventId, preview);
-      if (mounted) setState(() => _rootPreview = preview);
-    } catch (_) {}
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-    final tt = Theme.of(context).textTheme;
-    final threadRootId = widget.subGroup.threadRootId;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        if (threadRootId != null)
-          Padding(
-            padding: const EdgeInsets.fromLTRB(14, 6, 12, 2),
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Icon(Icons.forum_outlined, size: 14, color: cs.primary),
-                const SizedBox(width: 6),
-                Expanded(
-                  child: Text(
-                    _rootPreview ?? InboxText.inThread,
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                    style: tt.labelSmall?.copyWith(
-                      color: cs.primary,
-                      fontStyle: FontStyle.italic,
-                    ),
-                  ),
-                ),
-                const SizedBox(width: 6),
-                Text(
-                  InboxText.threadCount(widget.subGroup.notifications.length),
-                  style: tt.labelSmall?.copyWith(
-                    color: cs.onSurfaceVariant.withValues(alpha: 0.7),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        for (final notification in widget.subGroup.notifications)
-          _NotificationTile(
-            notification: notification,
-            client: widget.client,
-            threadRootId: threadRootId,
-          ),
-      ],
-    );
-  }
-}
-
-// ── Individual notification tile ─────────────────────────────
-class _NotificationTile extends StatelessWidget {
-  const _NotificationTile({
-    required this.notification,
-    required this.client,
-    required this.threadRootId,
-  });
-
-  final matrix_sdk.Notification notification;
-  final matrix_sdk.Client client;
-  final String? threadRootId;
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-    final tt = Theme.of(context).textTheme;
-    final event = notification.event;
-
-    final senderId = event.senderId;
-    final room = client.getRoomById(notification.roomId);
-    final senderName =
-        room?.unsafeGetUserFromMemoryOrFallback(senderId).calcDisplayname() ??
-            senderId;
-
-    final controller = context.read<InboxController>();
-    final body = _extractBody(context, event);
-    final ts = DateTime.fromMillisecondsSinceEpoch(notification.ts);
-    final isMention = controller.isMention(notification);
-
-    return InkWell(
-      mouseCursor: SystemMouseCursors.click,
-      canRequestFocus: false,
-      onTap: () {
-        final tid = threadRootId;
-        if (tid != null) {
-          context.goNamed(
-            Routes.roomThread,
-            pathParameters: {
-              RouteParams.roomId: notification.roomId,
-              RouteParams.eventId: tid,
-            },
-          );
-        } else {
-          context.goNamed(
-            Routes.room,
-            pathParameters: {RouteParams.roomId: notification.roomId},
-          );
-        }
-      },
-      child: Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // Unread indicator
-          if (!notification.read)
-            Padding(
-              padding: const EdgeInsets.only(top: 6, right: 6),
-              child: Container(
-                width: 8,
-                height: 8,
-                decoration: BoxDecoration(
-                  color: cs.primary,
-                  shape: BoxShape.circle,
-                ),
-              ),
-            )
-          else
-            const SizedBox(width: 14),
-
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  children: [
-                    Expanded(
-                      child: Text(
-                        senderName,
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                        style: tt.labelMedium?.copyWith(
-                          fontWeight: FontWeight.w600,
-                          color: cs.onSurface,
-                        ),
-                      ),
-                    ),
-                    if (isMention) ...[
-                      Icon(Icons.alternate_email_rounded,
-                          size: 12, color: cs.primary,),
-                      const SizedBox(width: 2),
-                      Text(
-                        threadRootId != null
-                            ? InboxText.mentionInThread
-                            : InboxText.mention,
-                        style: tt.labelSmall?.copyWith(
-                          color: cs.primary,
-                          fontSize: 10,
-                          fontWeight: FontWeight.w600,
-                        ),
-                      ),
-                      const SizedBox(width: 6),
-                    ],
-                    Text(
-                      formatRelativeTimestamp(ts),
-                      style: tt.bodySmall?.copyWith(
-                        fontSize: 11,
-                        color: cs.onSurfaceVariant.withValues(alpha: 0.5),
-                      ),
-                    ),
-                  ],
-                ),
-                if (body.isNotEmpty) ...[
-                  const SizedBox(height: 2),
-                  Text(
-                    body,
-                    maxLines: 2,
-                    overflow: TextOverflow.ellipsis,
-                    style: tt.bodyMedium?.copyWith(
-                      color: cs.onSurfaceVariant,
-                    ),
-                  ),
-                ],
-                const SizedBox(height: 6),
-              ],
-            ),
-          ),
-        ],
-      ),
-      ),
-    );
-  }
-
-  String _extractBody(BuildContext context, matrix_sdk.MatrixEvent event) {
-    final controller = context.read<InboxController>();
-    final content =
-        controller.decryptedContentFor(event.eventId) ?? event.content;
-    final msgtype = content['msgtype'];
-
-    if (msgtype == matrix_sdk.MessageTypes.Image) return InboxText.mediaImage;
-    if (msgtype == matrix_sdk.MessageTypes.Video) return InboxText.mediaVideo;
-    if (msgtype == matrix_sdk.MessageTypes.Audio) return InboxText.mediaAudio;
-    if (msgtype == matrix_sdk.MessageTypes.File) return InboxText.mediaFile;
-
-    final body = content['body'];
-    if (body is String) return stripReplyFallback(body);
-
-    return '';
-  }
-}
-
-// ── Invitations view ──────────────────────────────────────────
-class _InvitationsView extends StatelessWidget {
-  const _InvitationsView({required this.cs, required this.tt});
-
-  final ColorScheme cs;
-  final TextTheme tt;
-
-  @override
-  Widget build(BuildContext context) {
-    final selection = context.watch<SelectionService>();
-    final invitedRooms = selection.invitedRooms;
-    final invitedSpaces = selection.invitedSpaces;
-
-    if (invitedRooms.isEmpty && invitedSpaces.isEmpty) {
-      return Center(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Icon(Icons.mail_outline_rounded,
-                size: 56,
-                color: cs.onSurfaceVariant.withValues(alpha: 0.3),),
-            const SizedBox(height: 16),
-            Text(
-              InboxText.noPendingInvitations,
-              style: tt.titleMedium?.copyWith(
-                color: cs.onSurfaceVariant.withValues(alpha: 0.5),
-              ),
-            ),
-          ],
-        ),
-      );
-    }
-
-    return ListView(
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-      children: [
-        if (invitedSpaces.isNotEmpty) ...[
-          Padding(
-            padding: const EdgeInsets.fromLTRB(8, 8, 8, 4),
-            child: Text(
-              InboxText.sectionSpaces,
-              style: tt.labelLarge?.copyWith(
-                color: cs.onSurfaceVariant,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-          ),
-          for (final space in invitedSpaces) InviteTile(room: space),
-        ],
-        if (invitedRooms.isNotEmpty) ...[
-          Padding(
-            padding: const EdgeInsets.fromLTRB(8, 8, 8, 4),
-            child: Text(
-              InboxText.sectionRooms,
-              style: tt.labelLarge?.copyWith(
-                color: cs.onSurfaceVariant,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-          ),
-          for (final room in invitedRooms) InviteTile(room: room),
-        ],
-      ],
-    );
-  }
-}
-
-// ── Load more button ─────────────────────────────────────────
-class _LoadMoreButton extends StatelessWidget {
-  const _LoadMoreButton({
-    required this.isLoading,
-    required this.onPressed,
-  });
-
-  final bool isLoading;
-  final VoidCallback onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 8),
-      child: Center(
-        child: isLoading
-            ? const Padding(
-                padding: EdgeInsets.all(8),
-                child: SizedBox(
-                  width: 20,
-                  height: 20,
-                  child: CircularProgressIndicator(strokeWidth: 2),
-                ),
-              )
-            : TextButton(
-                onPressed: onPressed,
-                child: const Text(InboxText.loadMore),
-              ),
       ),
     );
   }
